+++ MPChartExample/src/com/xxmassdeveloper/mpchartexample/LineChartTime.java
@@ -95,7 +95,7 @@ public class LineChartTime extends DemoBase implements OnSeekBarChangeListener {
             private SimpleDateFormat mFormat = new SimpleDateFormat("dd MMM HH:mm");
 
             @Override
-            public String getFormattedValue(float value, AxisBase axis) {
+            public String getFormattedValue(double value, AxisBase axis) {
 
                 long millis = TimeUnit.HOURS.toMillis((long) value);
                 return mFormat.format(new Date(millis));
