+++ MPChartLib/src/main/java/com/github/mikephil/charting/renderer/AxisRenderer.java
@@ -115,7 +115,7 @@ public abstract class AxisRenderer extends Renderer {
      * @param min - the minimum value in the data object for this axis
      * @param max - the maximum value in the data object for this axis
      */
-    public void computeAxis(float min, float max, boolean inverted) {
+    public void computeAxis(double min, double max, boolean inverted) {
 
         // calculate the starting and entry point of the y-labels (depending on
         // zoom / contentrect bounds)
@@ -146,17 +146,17 @@ public abstract class AxisRenderer extends Renderer {
      *
      * @return
      */
-    protected void computeAxisValues(float min, float max) {
+    protected void computeAxisValues(double min, double max) {
 
-        float yMin = min;
-        float yMax = max;
+        double yMin = min;
+        double yMax = max;
 
         int labelCount = mAxis.getLabelCount();
         double range = Math.abs(yMax - yMin);
 
         if (labelCount == 0 || range <= 0 || Double.isInfinite(range)) {
-            mAxis.mEntries = new float[]{};
-            mAxis.mCenteredEntries = new float[]{};
+            mAxis.mEntries = new double[]{};
+            mAxis.mCenteredEntries = new double[]{};
             mAxis.mEntryCount = 0;
             return;
         }
@@ -189,10 +189,10 @@ public abstract class AxisRenderer extends Renderer {
 
             if (mAxis.mEntries.length < labelCount) {
                 // Ensure stops contains at least numStops elements.
-                mAxis.mEntries = new float[labelCount];
+                mAxis.mEntries = new double[labelCount];
             }
 
-            float v = min;
+            double v = min;
 
             for (int i = 0; i < labelCount; i++) {
                 mAxis.mEntries[i] = v;
@@ -224,7 +224,7 @@ public abstract class AxisRenderer extends Renderer {
 
             if (mAxis.mEntries.length < n) {
                 // Ensure stops contains at least numStops elements.
-                mAxis.mEntries = new float[n];
+                mAxis.mEntries = new double[n];
             }
 
             for (f = first, i = 0; i < n; f += interval, ++i) {
@@ -246,7 +246,7 @@ public abstract class AxisRenderer extends Renderer {
         if (mAxis.isCenterAxisLabelsEnabled()) {
 
             if (mAxis.mCenteredEntries.length < n) {
-                mAxis.mCenteredEntries = new float[n];
+                mAxis.mCenteredEntries = new double[n];
             }
 
             float offset = (float)interval / 2f;
